# backend/openapi.yaml
openapi: 3.0.3
info:
  title: RavelGo API
  version: "0.1.0"
  description: >
    Unified REST API for the RavelGo rider, driver, and admin apps.
    Authenticated with Amazon Cognito access tokens; role checks are enforced
    via Cognito group membership (`Rider`, `Driver`, `Admin`) — see
    `src/middleware/auth.ts`. There is also a WebSocket endpoint
    (`wss://<host>/ws?token=<access token>`) for live trip-status and
    driver-location push, not representable in OpenAPI 3.0 — see
    `docs/realtime-architecture.md` for its wire protocol. Every REST
    endpoint below has an HTTP-polling equivalent, so a client never
    strictly needs the WebSocket connection.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /trips/{id}/charge:
    post:
      tags: [Payments]
      summary: >-
        Charge the rider for a completed trip's final (tax-inclusive) fare
        (Driver on the trip, or Admin). CARD initializes a real Paystack
        transaction and returns the Payment as PENDING plus an
        authorizationUrl for the mobile app to open; the outcome lands
        asynchronously via POST /billing/webhook. WALLET debits the rider's
        Paystack-funded RavelGo balance and settles immediately. Cash is not
        a payment method.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                method: { type: string, enum: [CARD, WALLET], default: CARD }
      responses:
        "201":
          description: Payment record (PENDING with an authorizationUrl for CARD, SUCCEEDED for WALLET)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Payment"
                  - type: object
                    properties:
                      authorizationUrl: { type: string, nullable: true }
        "402":
          description: Wallet balance insufficient for the fare
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Trip not eligible to be charged, already charged, or rider has no wallet
  /payments/{id}/receipt:
    get:
      tags: [Payments]
      summary: A receipt for a charged trip (the paying rider, or Admin)
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Receipt
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentId: { type: string }
                  tripId: { type: string }
                  pickup: { type: string }
                  destination: { type: string }
                  amount: { type: number }
                  currency: { type: string }
                  method: { type: string, enum: [CARD, CASH, WALLET] }
                  status: { type: string, enum: [PENDING, SUCCEEDED, FAILED, REFUNDED] }
                  paidAt: { type: string, format: date-time, nullable: true }
                  issuedAt: { type: string, format: date-time }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /payments:
    get:
      tags: [Payments]
      summary: All payments (Admin)
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Paginated payments
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedPayment" }
...
  /trips/{id}/refund:
    post:
      tags: [Payments]
      summary: Refund a trip's payment, full or partial, reversing its commission (Admin, payouts:write)
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                amount: { type: number, description: Omit for a full refund }
                reason: { type: string }
      responses:
        "200":
          description: Refund result
          content:
            application/json:
              schema:
                type: object
                properties:
                  refunded: { type: number }
...
  /courier-requests/{id}/charge:
    post:
      tags: [Payments]
      summary: Charge the sender for a delivered package's final fare (Driver assigned to it, or Admin)
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                method: { type: string, enum: [CARD, WALLET, CASH] }
      responses:
        "201":
          description: Payment
  /courier-requests/{id}/pay:
    post:
      tags: [Payments]
      summary: Pay for my own delivered package (Rider who sent it)
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                method: { type: string, enum: [CARD, WALLET, CASH] }
      responses:
        "201":
          description: Payment
  /courier-requests/{id}/refund:
    post:
      tags: [Payments]
      summary: Refund a delivery's payment, full or partial (Admin, payouts:write)
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                amount: { type: number }
                reason: { type: string }
      responses:
        "200":
          description: Refund result
...
  /trips/{id}/pay:
    post:
      tags: [Payments]
      summary: Pay for my completed trip by CARD, CASH or WALLET (Rider)
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [method]
              properties:
                method: { type: string }
      responses:
        "201":
          description: Payment
          content:
            application/json:
              schema: { type: object }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Cognito access token
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    BadRequest:
      description: Validation or bad request error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Caller lacks the required Cognito group
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    Payment:
      type: object
      properties:
        id: { type: string }
        tripId: { type: string }
        userId: { type: string }
        amount: { type: number }
        currency: { type: string }
        method: { type: string, enum: [CARD, CASH, WALLET] }
        status: { type: string, enum: [PENDING, SUCCEEDED, FAILED, REFUNDED] }
...
    PaginatedPayment:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Payment" } }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }

    # Standardized error response (PL-01)
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, timestamp]
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - VALIDATION_ERROR
                - AUTHENTICATION_ERROR
                - AUTHORIZATION_ERROR
                - NOT_FOUND
                - CONFLICT
                - INTERNAL_SERVER_ERROR
                - RATE_LIMIT_EXCEEDED
                - BAD_REQUEST
                - PAYMENT_ERROR
                - DATABASE_ERROR
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              nullable: true
              description: Field-level validation errors
              items:
                type: object
                properties:
                  field: { type: string, description: 'Field name or path (e.g., "email", "address.zipCode")' }
                  message: { type: string, description: Error message for this field }
                  code: { type: string, nullable: true, description: 'Error code (e.g., "too_small", "invalid_email")' }
            timestamp:
              type: string
              format: date-time
              description: When the error occurred (ISO 8601)